+++ out/extension.js
@@ -19,6 +19,9 @@ function activate(context) {
     vscode.commands.registerCommand('aws-access-vscode-extension.RunLoginCommand', () => {
         StatusBar.StatusBarItem.Current.RunLoginCommand();
     });
+    vscode.commands.registerCommand('aws-access-vscode-extension.PauseAutoLogin', () => {
+        StatusBar.StatusBarItem.Current.PauseAutoLogin();
+    });
     vscode.commands.registerCommand('aws-access-vscode-extension.SetActiveProfile', () => {
         StatusBar.StatusBarItem.Current.SetActiveProfile();
     });