+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+# v1.0.4
+    - Pause/UnPause Auto Login
+    - Show Expire Time in the tooltip
 # v0.0.1
     -initial release